refactor: read bytes to structs via memory layout instead of manually

+++ src/DIBHeader.hpp
@@ -1,17 +1,18 @@
+#include <cstdint>
 #include <ostream>
 
 
 struct DIBHeader {
-    int dibHeaderSize;
-    int width;
-    int height;
-    int colorPlanes;
-    int bitsPerPixel;
-    int compressionMethod;
-    int imageSize;
-    int horizontalRes;
-    int verticalRes;
-    int nColors;
-    int nImportantColors;
+    uint32_t dibHeaderSize;
+    uint32_t width;
+    uint32_t height;
+    uint16_t colorPlanes;
+    uint16_t bitsPerPixel;
+    uint32_t compressionMethod;
+    uint32_t imageSize;
+    uint32_t horizontalRes;
+    uint32_t verticalRes;
+    uint32_t nColors;
+    uint32_t nImportantColors;
 };
 std::ostream& operator<<(std::ostream& os, const DIBHeader& dibHeader);
